Remove Commented Code and Refactor Dispatcher
